requirements.txt
colorama==0.4.6
greenlet==3.3.2
iniconfig==2.3.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.48
tenacity==9.1.4
typing_extensions==4.15.0
tzdata==2025.3
